+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Alt-Left and Alt-Right for history navigation (#23)
+
 ### Fixed
 - Actual unicode bullet symbol is used for lists: U+2022
 - Performance when loading very long cached pages improved (#26)